refactor: move components in HomePage to ./sections/ folder
