+++ packages/static-ui/src/containers/Home/section/aboutUs.jsx
