+++ packages/static-ui/src/containers/Home/section/event.jsx
