+++ packages/static-ui/src/containers/Home/section/footer.jsx
