+++ packages/static-ui/src/containers/Home/section/hero.jsx
